Only handle newlines as separate parses when in lore mode
Currently, the backend splits all newlines and parses the split-up
messages with completely separate MiniMessage instances so they have
separate parents. With this change, this will be only done when in lore
mode, and newlines will simply remain in the message regularly in all
other modes.

The justification for this is the fact that newlines can't appear in
lores at all, as the client will render them in a tofubox. Therefore
having newlines in the lore mode in the webui implies the fact that
these are in fact separate lines with separate tree roots.

This isn't explained on the site though, maybe it should be...