Listen for when the websocket closes and display a warning
This shouldn't normally happen, but if it does, previous the input
would simply stop working. With this change, it displays a warning

It could also attempt to reconnect or something, but given that a
websocket closure should *only* happen in weird cases, such as losing
all internet connection, I have low confidence in a reconnect being
successful at all. The user can simply manually do a reconnect by
using the refresh button anyway