feat: downsampling setting
This adds a new option to the settings which allows for setting a
downsampler, which is run before converting the component to HTML.

This allows for "simulating" either legacy, or bedrock mode.

Has a TODO to resolve, will probably need better visual clarity (i.e.
something on the main page to show that you're currently in downsample
mode, so you don't forget...?)
4.14.0-SNAPSHOT is currently required since it's the first version which
makes `TextColor.nearestColorTo` public.

Closes #153